requirements.txt
fastapi
uvicorn
chromadb
openai
python-dotenv
pydantic
numpy
scikit-learn
requests
sentence-transformers